requirements.txt
# Core ML & Computer Vision
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.3

# Data Processing & Analysis
pandas>=2.0.3
scikit-learn>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0

# Data Augmentation
albumentations>=1.3.1

# Model Interpretability
pytorch-grad-cam>=1.4.6

# Web UI
gradio>=4.0.0

# Configuration & Utilities
pyyaml>=6.0.1
tqdm>=4.66.1

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Optimization (Optional)
onnxruntime>=1.16.0